Made it so that Makkusu made the monitoring tech
Makkusu created the panopticon like tech to create transparency. This enforces his character as he believes in telling nothing but the truth to a fault.

Spike on the other hand wants a world with false transparency so that he can create the narrative for the world not for peace but profits. It just so happens that he eventually see peace as potentially more profitable than war.

Changed so that Makkusu alone is making a statement about the war

Added  epilogue easter egg revealing that his monologue at the beginning of the story was him logging the beginning of an entry for future generations to view