Switched to Tomasia being the mentor
This gives her a bigger role whereas before it was none existent and would randomly appear at the end of the story. It better fits her character to be his mentor as opposed Kyshumu or even Xenos. It also means that I don't need to bring back Xenos which would prevent the use of certain maybe important and emotional plot points.