Change main character
Changed the protagonist from Max to Rae and made Max into Rae's father.

This also meant changing the name of the folder from max to cyber.fighter. Using a more generic name rather than a characters  name makes it less likely the folder name will need to be changed again in the future if the character name is changed again.