Fixed Typos and expanded scenes & dialouge
Expanded on dialouge for Rae asking for a bigger role on the team. It shows where Rae is in her charcter arc

Provided an explanation as to why Aarock was beat by Rae so easily. This provides character depth for Aarock.

Rae provides exposition on what it is like to fuse with a cyber.beast