Added potential scenes
Added scene in which Max points out Airiona's hipocracy in her actions which will allow her to turn a new leaf

Added scene to show Max using Minia cyber.beast in what may or may not be his first fight. This will foreshadow Torrey and Mina's origin in regards to their relation with Max.

added what may be a plot beat for subplot of Minia and Airiona's shaky relationship