Remove event, add other scences & dialouge
Remove the "never ending" match as it conflicts the rules of the CNS and the world.  It is the rule by which matches are won after the data which is limited in satilites are corupted not when cybervators run out of energy. 3-5 Cybervators should be enough to complete this task

Scene with Minia and Max will show their growing relationship

Added possible explaination for how Minia became to be

The other lines of diaouge will serve as exposition to explain cyber.beasts